# src/main/resources/openapi.yaml
openapi: 3.1.1
info:
  title: Distributed Messenger API
  version: 1.0.0
servers:
- description: API Version 2
  url: /api/v2
security:
- bearerAuth: []
...
paths:
...
  /chats/{chatId}/messages:
    get:
      operationId: getChatMessages
      parameters:
      - explode: false
        in: path
        name: chatId
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      - explode: true
        in: query
        name: limit
        required: false
        schema:
          default: 50
          type: integer
        style: form
      - description: Курсор для получения сообщений старше указанной временной метки.
        explode: true
        in: query
        name: before_timestamp
        required: false
        schema:
          format: date-time
          type: string
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedMessages"
          description: Страница со списком сообщений.
        "404":
          description: Чат не найден.
      summary: Получить историю сообщений
      tags:
      - Messages
    post:
      operationId: sendMessage
      parameters:
      - explode: false
        in: path
        name: chatId
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MessageCreate"
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
          description: Сообщение создано.
        "404":
          description: Чат не найден.
      summary: Отправить сообщение
      tags:
      - Messages
  /chats/{chatId}/messages/{messageId}:
    delete:
      operationId: deleteMessage
      parameters:
      - explode: false
        in: path
        name: chatId
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      - explode: false
        in: path
        name: messageId
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      responses:
        "204":
          description: Сообщение удалено.
        "403":
          description: Доступ запрещен.
        "404":
          description: Сообщение не найдено.
      summary: Удалить сообщение
      tags:
      - Messages
    get:
      operationId: getMessageById
      parameters:
      - explode: false
        in: path
        name: chatId
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      - explode: false
        in: path
        name: messageId
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
          description: Данные сообщения.
        "404":
          description: Сообщение не найдено.
      summary: Получить сообщение по ID
      tags:
      - Messages
    patch:
      operationId: editMessage
      parameters:
      - explode: false
        in: path
        name: chatId
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      - explode: false
        in: path
        name: messageId
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MessageUpdate"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
          description: Сообщение обновлено.
        "403":
          description: Доступ запрещен.
        "404":
          description: Сообщение не найдено.
      summary: Редактировать сообщение
      tags:
      - Messages
  /chats/{chatId}/messages/{messageId}/history:
    get:
      operationId: getMessageHistory
      parameters:
      - explode: false
        in: path
        name: chatId
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      - explode: false
        in: path
        name: messageId
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/MessageHistory"
                type: array
          description: Список изменений.
        "404":
          description: Сообщение не найдено.
      summary: Получить историю изменений сообщения
      tags:
      - Messages
...
  /users/{userId}/settings:
    get:
      operationId: getAppSettings
      parameters:
      - explode: false
        in: path
        name: userId
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppSettings"
          description: Объект с настройками.
        "403":
          description: Доступ запрещен.
      summary: Получить настройки пользователя
      tags:
      - Settings
    patch:
      operationId: updateAppSettings
      parameters:
      - explode: false
        in: path
        name: userId
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AppSettings"
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppSettings"
          description: Настройки обновлены.
        "403":
          description: Доступ запрещен.
      summary: Обновить настройки пользователя
      tags:
      - Settings
components:
  schemas:
...
    Message:
      example:
        senderId: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        chatId: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        content: content
        fileId: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        timestamp: 2000-01-23T04:56:07.000+00:00
      properties:
        id:
          format: uuid
          type: string
        senderId:
          format: uuid
          type: string
        chatId:
          format: uuid
          type: string
        content:
          type: string
        fileId:
          format: uuid
          type: string
          nullable: true
        timestamp:
          format: date-time
          type: string
      required:
      - chatId
      - content
      - id
      - senderId
      - timestamp
    MessageHistory:
      example:
        historyId: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        messageId: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        editTimestamp: 2000-01-23T04:56:07.000+00:00
        editedContent: editedContent
      properties:
        historyId:
          format: uuid
          type: string
        messageId:
          format: uuid
          type: string
        editedContent:
          type: string
        editTimestamp:
          format: date-time
          type: string
      required:
      - editTimestamp
      - editedContent
      - historyId
      - messageId
    MessageCreate:
      example:
        content: content
        fileId: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
      properties:
        content:
          type: string
        fileId:
          format: uuid
          type: string
          nullable: true
      required:
      - content
    MessageUpdate:
      example:
        content: content
      properties:
        content:
          type: string
      required:
      - content
...
    AppSettings:
      example:
        MAX_FILE_SIZE: 6
        THEME: 0
        HISTORY_STORAGE: 5
        MESSAGE_ENCRYPTION: 1
      properties:
        THEME:
          description: "0: System, 1: Light, 2: Dark"
          enum:
          - 0
          - 1
          - 2
          type: integer
        MAX_FILE_SIZE:
          enum:
          - 10
          - 50
          - 100
          type: integer
        MESSAGE_ENCRYPTION:
          enum:
          - 0
          - 1
          type: integer
        HISTORY_STORAGE:
          enum:
          - 7
          - 30
          - 365
          type: integer
...
    PaginatedMessages:
      description: "Страница с сообщениями, использующая курсорную пагинацию."
      example:
        nextCursor: 2000-01-23T04:56:07.000+00:00
        hasNextPage: true
        items:
        - senderId: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          chatId: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          content: content
          fileId: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          timestamp: 2000-01-23T04:56:07.000+00:00
        - senderId: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          chatId: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          content: content
          fileId: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          timestamp: 2000-01-23T04:56:07.000+00:00
      properties:
        items:
          items:
            $ref: "#/components/schemas/Message"
          type: array
        nextCursor:
          description: Timestamp последнего сообщения для запроса следующей страницы.
          format: date-time
          type: string
          nullable: true
        hasNextPage:
          type: boolean
...
  securitySchemes:
    bearerAuth:
      bearerFormat: JWT
      scheme: bearer
      type: http